More permissive datetime parsing
When chrono_time is enabled, the datetime parser is now more
permissive and will also accept a datetime format where the seconds
are missing (like in the example from the PDF spec), and also just
a date without any time information. Defaults are used for the
missing fields, as required by the PDF spec.