Genericize Content to allow AsRef<[Operation]>
This was just Vec<Operation> before, which may be the most common
use-case and is also required for parsing, since allocation is
necessary.
When creating a document allocating a Vec per Content may not be
necessary though, so for that use-case allowing any AsRef<[Operation]>,
like just a slice, is convenient.