Guard example based on if the "parser" feature is enabled
Now that the examples are being run as part of the CI one of the CI
tests checks if the code compiles with no default features. Since the
example is dependant on one of the parsers being enabled adjust the
code to check for the parser feature.

This fixes the CI build error.

Signed-off-by: Wiktor Kwapisiewicz <wiktor@metacode.biz>