Handle references to arrays in get_page_contents() (#245)
Currently we assume that we if get a reference it's a reference to a
stream. However, it can also be a reference to an array as is the case
in https://github.com/jrmuizel/pdf-extract/files/12017696/Model-Card-Claude-2.pdf

Handle this by checking if we get a stream and otherwise dereferencing
it.

This should also handle the case where there's a reference to a
reference to a stream.